refactor: 투표 도메인 폴더 구조 및 통계 집계 로직 리팩토링 (#388)
- 투표 통계 자동 집계 로직을 VoteService 내부로 이동
- 불필요한 VoteStatisticsScheduler 클래스 제거
- VoteStatistics 엔티티 및 관련 클래스 제거
