+++ backend/src/main/java/com/tamnara/backend/poll/service/PollServiceImpl.java
@@ -6,12 +6,16 @@ import com.tamnara.backend.alarm.event.AlarmEvent;
 import com.tamnara.backend.poll.domain.Poll;
 import com.tamnara.backend.poll.domain.PollOption;
 import com.tamnara.backend.poll.domain.PollState;
+import com.tamnara.backend.poll.domain.VoteStatistics;
+import com.tamnara.backend.poll.dto.OptionResult;
 import com.tamnara.backend.poll.dto.PollInfoDTO;
 import com.tamnara.backend.poll.dto.request.PollCreateRequest;
 import com.tamnara.backend.poll.dto.response.PollInfoResponse;
+import com.tamnara.backend.poll.dto.response.PollStatisticsResponse;
 import com.tamnara.backend.poll.repository.PollOptionRepository;
 import com.tamnara.backend.poll.repository.PollRepository;
 import com.tamnara.backend.poll.repository.VoteRepository;
+import com.tamnara.backend.poll.repository.VoteStatisticsRepository;
 import com.tamnara.backend.user.domain.User;
 import com.tamnara.backend.user.repository.UserRepository;
 import jakarta.transaction.Transactional;
@@ -41,8 +45,9 @@ public class PollServiceImpl implements PollService {
 
     private final PollRepository pollRepository;
     private final PollOptionRepository pollOptionRepository;
-    private final UserRepository userRepository;
     private final VoteRepository voteRepository;
+    private final VoteStatisticsRepository voteStatisticsRepository;
+    private final UserRepository userRepository;
 
     @Transactional
     public Long createPoll(PollCreateRequest request) {
@@ -112,6 +117,23 @@ public class PollServiceImpl implements PollService {
         } else {
             log.warn("[WARN] 투표 삭제 대상 없음 - 공개 중인 투표가 존재하지 않음");
         }
+    }
+
+    public PollStatisticsResponse getPollStatistics(Long pollId) {
+        Poll poll = pollRepository.findById(pollId)
+                .orElseThrow(() -> new ResponseStatusException(HttpStatus.NOT_FOUND, POLL_NOT_FOUND));
+
+        List<PollOption> options = poll.getOptions();
+
+        List<OptionResult> results = options.stream().map(option -> {
+            VoteStatistics stat = voteStatisticsRepository.findByPollIdAndOptionId(pollId, option.getId())
+                    .orElse(VoteStatistics.zero(poll, option));
+            return new OptionResult(option.getId(), option.getTitle(), stat.getCount());
+        }).collect(Collectors.toList());
+
+        long totalVotes = results.stream().mapToLong(OptionResult::getCount).sum();
+
+        return new PollStatisticsResponse(pollId, results, totalVotes);
     }
