--- backend/src/main/java/com/tamnara/backend/poll/service/VoteStatisticsService.java
@@ -1,9 +1,0 @@
-package com.tamnara.backend.poll.service;
-
-import com.tamnara.backend.poll.dto.response.PollStatisticsResponse;
-
-public interface VoteStatisticsService {
-    void generateAllStatistics();
-    PollStatisticsResponse getPollStatistics(Long pollId);
-
-}
