+++ backend/src/test/java/com/tamnara/backend/poll/config/PollTestConfig.java
@@ -2,7 +2,6 @@ package com.tamnara.backend.poll.config;
 
 import com.tamnara.backend.poll.service.PollService;
 import com.tamnara.backend.poll.service.VoteService;
-import com.tamnara.backend.poll.service.VoteStatisticsService;
 import org.mockito.Mockito;
 import org.springframework.boot.test.context.TestConfiguration;
 import org.springframework.context.annotation.Bean;
@@ -18,10 +17,5 @@ public class PollTestConfig {
     @Bean
     public VoteService voteService() {
         return Mockito.mock(VoteService.class);
-    }
-
-    @Bean
-    public VoteStatisticsService voteStatisticsService() {
-        return Mockito.mock(VoteStatisticsService.class);
     }
 }
